web-app/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
Authlib==1.3.0
azure-common==1.1.28
azure-core==1.29.5
azure-identity==1.16.0
azure-keyvault-secrets==4.8.0
azure-mgmt-core==1.4.0
azure-mgmt-datafactory==8.0.0
azure-search-documents==11.4.0
azure-storage-blob==12.19.0
blinker==1.8.2
cachelib==0.13.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
dataclasses-json==0.6.4
distro==1.8.0
Django==5.0.4
django-formtools==2.5.1
django-otp==1.5.1
django-phonenumber-field==7.3.0
django-two-factor-auth==1.16.0
exceptiongroup==1.2.1
fastapi==0.110.3
Flask==3.0.3
Flask-Session==0.8.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
httpx-sse==0.4.0
idna==3.7
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.1.14
langchain-cli==0.0.21
langchain-community==0.0.31
langchain-core==0.1.39
langchain-experimental==0.0.56
langchain-text-splitters==0.0.1
langserve==0.0.51
langsmith==0.1.39
lxml==5.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
mdurl==0.1.2
msal==1.28.0
msal-extensions==1.1.0
msgspec==0.18.6
mssql-django==1.5
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
oauthlib==3.2.2
openai==1.3.5
orjson==3.10.4
packaging==23.2
phonenumbers==8.13.42
portalocker==2.8.2
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
PyJWT==2.8.0
pyodbc==5.0.1
pypdf==3.17.1
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.0
pytz==2024.1
pyxero==0.9.4
PyYAML==6.0.1
qrcode==7.4.2
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.1
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.29
sqlparse==0.5.0
sse-starlette==1.8.2
starlette==0.37.2
stripe==5.5.0
tenacity==8.3.0
tomlkit==0.12.5
tqdm==4.66.4
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
uuid==1.30
uvicorn==0.23.2
Werkzeug==3.0.3
xero_python==3.0.0
yarl==1.9.4
